examples: Switched to ed25519, nodes -> noenc
I switched the curve to ed25519, a generally more trustworthy curve and the default in openssh.
As a bonus, it *really* simplifies what is already a complicated command.
The old command wouldn't even run in all shells because it used process substitution.

'nodes' is deprecated in favour of 'noenc', so I switched that too.